% Auto-generated: do not edit by hand
\name{'vtk'ShareDataSet}

\alias{'vtk'ShareDataSet}

\title{ShareDataSet component}

\description{
ShareDataSet capture a dataset or a source and allow it to use it in another pipeline or representation.
}

\usage{
'vtk'ShareDataSet(children=NULL, id=NULL, name=NULL, port=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }

\item{id}{Character. The ID used to identify this component.}

\item{name}{Character. Unique dataset name to cross reference}

\item{port}{Numeric. downstream connection port}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

