% Auto-generated: do not edit by hand
\name{vtkCellData}

\alias{vtkCellData}

\title{CellData component}

\description{
CellData is exposing a vtkCellData to a downstream element
}

\usage{
vtkCellData(children=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

