% Auto-generated: do not edit by hand
\name{vtkFieldData}

\alias{vtkFieldData}

\title{FieldData component}

\description{
FieldData is exposing a FieldData to a downstream element
}

\usage{
vtkFieldData(children=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

