% Auto-generated: do not edit by hand
\name{vtkPointData}

\alias{vtkPointData}

\title{PointData component}

\description{
PointData is exposing a vtkPointData to a downstream element
}

\usage{
vtkPointData(children=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

