% Auto-generated: do not edit by hand
\name{vtkVolumeController}

\alias{vtkVolumeController}

\title{VolumeController component}

\description{
VolumeController is a GUI to control the piecewise function
}

\usage{
vtkVolumeController(id=NULL, rescaleColorMap=NULL, size=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component.}

\item{rescaleColorMap}{Logical. Use opacity range to rescale color map}

\item{size}{List of numerics. Controller size in pixels}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

