% Auto-generated: do not edit by hand
\name{vtkVtkAlgorithm}

\alias{vtkVtkAlgorithm}

\title{VtkAlgorithm component}

\description{
VtkAlgorithm is exposing a source or filter to a downstream filter It takes the following set of properties:   - vtkClass: vtkClassName   - state: {}
}

\usage{
vtkVtkAlgorithm(children=NULL, id=NULL, port=NULL, vtkClass=NULL, state=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{port}{Numeric. downstream connection port}

\item{vtkClass}{Character. vtkClass name}

\item{state}{Named list. set of property values for vtkClass}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

