% Auto-generated: do not edit by hand
\name{vtkVtkCellData}

\alias{vtkVtkCellData}

\title{VtkCellData component}

\description{
VtkCellData is exposing a VtkCellData to a downstream element It provides the following properties to its children:   - pass along: 'view', 'representation`, `setProps`   - `fields` == `VtkCellData`
}

\usage{
vtkVtkCellData(children=NULL, id=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. List of representation to show}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

