% Auto-generated: do not edit by hand
\name{vtkVtkView}

\alias{vtkVtkView}

\title{VtkView component}

\description{
VtkView is responsible to render vtk.js data. It takes the following set of properties:   - `background`:
}

\usage{
vtkVtkView(children=NULL, id=NULL, background=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. List of representation to show}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{background}{Unnamed list. The color of the view background using 3 floating numbers
between 0-1 of Red, Green, Blue component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

