% Auto-generated: do not edit by hand
\name{vtkAlgorithm}

\alias{vtkAlgorithm}

\title{Algorithm component}

\description{
Algorithm is exposing a source or filter to a downstream filter It takes the following set of properties:   - vtkClass: vtkClassName   - state: {}
}

\usage{
vtkAlgorithm(children=NULL, id=NULL, port=NULL, state=NULL, vtkClass=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }

\item{id}{Character. The ID used to identify this component.}

\item{port}{Numeric. downstream connection port}

\item{state}{Named list. set of property values for vtkClass}

\item{vtkClass}{Character. vtkClass name}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

