% Auto-generated: do not edit by hand
\name{vtkShareDataSet}

\alias{vtkShareDataSet}

\title{ShareDataSet component}

\description{
ShareDataSet capture a dataset or a source and allow it to use it in another pipeline or representation.
}

\usage{
vtkShareDataSet(children=NULL, id=NULL, port=NULL, name=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }

\item{id}{Character. The ID used to identify this component.}

\item{port}{Numeric. downstream connection port}

\item{name}{Character. Unique dataset name to cross reference}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

