% Auto-generated: do not edit by hand
\name{vtkVtkCellData}

\alias{vtkVtkCellData}

\title{VtkCellData component}

\description{
VtkCellData is exposing a vtkCellData to a downstream element
}

\usage{
vtkVtkCellData(children=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. List of representation to show}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

