% Auto-generated: do not edit by hand
\name{vtkVtkFieldData}

\alias{vtkVtkFieldData}

\title{VtkFieldData component}

\description{
VtkFieldData is exposing a vtkFieldData to a downstream element
}

\usage{
vtkVtkFieldData(children=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. List of representation to show}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

