% Auto-generated: do not edit by hand
\name{vtkVtkPointData}

\alias{vtkVtkPointData}

\title{VtkPointData component}

\description{
VtkPointData is exposing a vtkPointData to a downstream element
}

\usage{
vtkVtkPointData(children=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. List of representation to show}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

